requirements.txt
fastapi
uvicorn
aiogram
python-dotenv
selenium
webdriver-manager
nest-asyncio
aiohttp
pydantic